+++ app/src/pages/SingleFund/SingleFund.js
@@ -1,9 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { withRouter } from 'react-router-dom';
-import { Form } from '../../components/Form/Form';
-import { funderInputs } from '../../components/Form/inputs';
 import ContributionForm from './subcomponents/ContributeForm';
 import ProgressBar from './subcomponents/ProgressBar';
+import { UpdateFunderForm } from './subcomponents/UpdateFunderForm';
 
 const SingleFund = ({ drizzle, drizzleState, match }) => {
   const [state, setState] = useState({
@@ -11,35 +10,12 @@ const SingleFund = ({ drizzle, drizzleState, match }) => {
   });
   const { funderId } = match.params;
 
-  const [fund, setFund] = useState({
-    title: '',
-    goal: '',
-    endDate: '',
-    description: '',
-  });
-
   useEffect(() => {
     const funderHash = drizzle.contracts.PleaseFundMe_v3.methods.getFunderById.cacheCall(
       funderId,
     );
     setState({ funderHash });
   }, [funderId, drizzle.contracts.PleaseFundMe_v3.methods.getFunderById]);
-
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-    setFund((prevState) => ({ ...prevState, [name]: value }));
-  };
-
-  const updateFunder = async (event) => {
-    event.preventDefault();
-    drizzle.contracts.PleaseFundMe_v3.methods.updateFunder.cacheSend(
-      funderId,
-      fund.title,
-      fund.goal,
-      fund.description,
-      123456,
-    );
-  };
 
   const withdrawFunder = () => {
     drizzle.contracts.PleaseFundMe_v3.methods.withdraw.cacheSend(funderId);
@@ -59,20 +35,19 @@ const SingleFund = ({ drizzle, drizzleState, match }) => {
         <p>Fund Balance: {funder.fundBalance}</p>
         <p>Total Amount Raised: {funder.amountRaised}</p>
         <p>Description: {funder.description}</p>
-        <p>End Date: {funder.endDate}</p>
+        <p>Start Date: {new Date(funder.startDate * 1000).toDateString()}</p>
+        <p>End Date: {new Date(funder.endDate * 1000).toDateString()}</p>
+        <p>Last Updated: {new Date(funder.updatedAt * 1000).toDateString()}</p>
       </div>
       <div className="side-bar">
         <ProgressBar amount={funder.amountRaised} goal={funder.fundTarget} />
       </div>
       {isOwner && (
-        <div className="fund-info">
-          <Form
-            inputs={funderInputs}
-            values={fund}
-            handleChange={handleChange}
-            onSubmit={updateFunder}
-          />
-        </div>
+        <UpdateFunderForm
+          drizzle={drizzle}
+          funderId={funder.id}
+          initialState={funder}
+        />
       )}
       {isOwner ? (
         <div className="side-bar">
